* Immutable AST

This  PR  introduces a  secondary  data  model  for  the org-mode  AST:  an
persistent tree  that can  efficiently store all  versions of  the modified
data. This data  model is intended to be used  for various GUI applications
and other places where  it is necessary to edit the tree  and maybe undo or
redo some things.

Immutable AST is not  a drop-in replacement of the regular  sem AST -- most
of  the  existing  code  (built-in exporters,  python  exporters,  parsing,
protobuf serialization, tree  repr debugging) works only with  the sem AST,
and it is necessary  to convert the immutable AST to sem  and vice versa to
perform operations.

Immutable  AST  tree structure,  on  the  other  hand,  is intended  to  be
identical to the sem  AST, save for the subnode access and  most of the AST
types. Enums are shared, as well as simple built-in structures that I could
move out of the type definition.

Immutable AST idea is built on top of the earlier implementation of the sem
AST -- store all the node values in a  flat chunk of data and use DOD ID to
access the  results. It did not  work out because it  is so hard to  do the
tree mutation in  this context -- basically it is  impossible to remove the
old data from  the store, and there are lots  of annoying relocations. Plus
python API  becomes more complex as  well. So I originally  moved from this
idea back then, but with the new immutable AST these disadvantages actually
became features. Not being able to remove old versions of the nodes -- this
part allowed me to build the  "store all versions of the document". Editing
complexities were  more difficult to  fix, but with  the right API  it also
worked out in the end.

In addition to  the immutable AST data  model I also re-added  the mind map
implementation based on the immutable AST IDs.

* UI changes

This PR  removes Qt example  -- I realized  that I'm extremely  unlikely to
finish  a full-featured  GUI  application using  Qt in  the  scope of  this
project, immutable  AST makes it easy  to work with the  immediate mode GUI
applications, and imgui provides a /much/ simpler development experience to
quickly arrange  an example interactive application  without getting bogged
too deeply into the implementation details.  So I cut the Qt implementation
and decided to go for the

+++ scripts/cxx_repository/code_forensics.cpp
@@ -61,7 +61,7 @@ void InsertCommits(
 
     for (const auto& [id, item] : commits.pairs()) {
         query.bind(1, idcast(id.getValue()));
-        query.bind(2, item->author.getValue());
+        query.bind(2, static_cast<i64>(item->author.getValue()));
         query.bind(3, static_cast<i64>(item->time));
         query.bind(4, item->timezone);
         query.bind(5, item->hash.toBase());
@@ -164,7 +164,7 @@ void InsertAuthors(
     SQLite::Statement query(db, "INSERT INTO Author VALUES (?, ?, ?)");
 
     for (const auto& [id, item] : authors.pairs()) {
-        query.bind(1, id.getValue());
+        query.bind(1, static_cast<i64>(id.getValue()));
         query.bind(2, item->name.toBase());
         query.bind(3, item->email.toBase());
